project/requirements.txt
# ── 正式環境套件（production）─────────────────────────────────────────────────
Django>=3.2,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5
psycopg2-binary>=2.9.0
